+++ models/Exam.ts
@@ -1,0 +1,11 @@
+
+class Exam {
+    constructor(
+        public key: number,
+        public id: string,
+        public type: string,
+        public scoreScale: number[]
+    ) {}
+}
+
+export default Exam
