+++ models/Section.ts
@@ -1,12 +1,12 @@
-import ProblemModel from "./Problem";
+// import ProblemModel from "./Problem";
 
-class Section {
-    constructor(
-        public key: number,
-        public id: string,
-        public parentId: string,
-        public problems: ProblemModel[]
-    ) {}
-}
+// class Section {
+//     constructor(
+//         public key: number,
+//         public id: string,
+//         public parentId: string,
+//         public problems: ProblemModel[]
+//     ) {}
+// }
 
-export default Section;
+// export default Section;
