--- models/Test.js
@@ -1,11 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-class Student {
-    constructor(key, fName, lName, pastWork) {
-        this.key = key;
-        this.fName = fName;
-        this.lName = lName;
-        this.pastWork = pastWork;
-    }
-}
-exports.default = Student;
