+++ package.json
@@ -10,8 +10,14 @@
     "test": "jest"
   },
   "dependencies": {
+    "@prisma/client": "^5.11.0",
+    "@types/node": "^20.11.26",
+    "@types/validator": "^13.11.9",
+    "Express": "^3.0.1",
     "jest": "^29.7.0",
-    "sequelize": "^6.37.1",
+    "prisma": "^5.11.0",
+    "reflect-metadata": "^0.2.1",
+    "sequelize-typescript": "^2.1.6",
     "sqlite3": "^5.1.7",
     "typescript": "^5.3.3"
   },
